refactor: refactored statistics, moved about page to separate file
